chore: console.log 삭제
