+++ app/(route)/_providers/index.ts
@@ -4,4 +4,3 @@ export { default as AutoLoginPresence } from './AutoLoginPresence';
 export { default as RecoilRootWrapper } from './RecoilRootWrapper';
 export { default as QueryWrapper } from './QueryWrapper';
 export { default as PopupWrapper } from './PopupWrapper';
-export { default as SyncTheme } from './SyncTheme';
