+++ app/(route)/layout.tsx
@@ -8,37 +8,12 @@ import {
   RecoilRootWrapper,
 } from 'app/(route)/_providers';
 import { DefaultLayout } from 'app/_components/layout';
-import Script from 'next/script';
+import SyncThemeScript from './_providers/SyncThemeScript';
 
 export const metadata = {
   title: 'Casper',
   description: 'Changwon National University Casper',
 };
-
-function SyncThemeScript() {
-  const scriptFn = () => {
-    const savedTheme = localStorage.getItem('theme');
-
-    const htmlElement = document.querySelector('html');
-
-    switch (savedTheme) {
-      case 'ligth':
-        htmlElement?.classList.remove('dark');
-        break;
-      case 'dark':
-        htmlElement?.classList.add('dark');
-        break;
-      default:
-        htmlElement?.classList.add('dark');
-    }
-  };
-
-  const scriptStr = `
-    (${scriptFn.toString()})();
-  `;
-
-  return <Script id="theme" dangerouslySetInnerHTML={{ __html: scriptStr }} />;
-}
 
 export default function RootLayout({ children }: { children: ReactNode }) {
   return (
