--- app/_utils/syncTheme.ts
@@ -1,3 +1,0 @@
-(function initTheme() {
-  console.log('test');
-})();
